+++ example/TestLinkExampleGenericApi.py
@@ -101,6 +101,7 @@ myTLVersionShort = myTLVersion.replace('.', '')
 
 NEWPROJECT="PROJECT_API_GENERIC-%s" % myPyVersionShort
 NEWPREFIX="GPROAPI%s" % myPyVersionShort
+ITSNAME="myITS"
 
 # used connection settings
 print(myTestLink.connectionInfo())
@@ -146,11 +147,16 @@ try:
 except TLResponseError:
     print("No project with prefix %s exists" % NEWPREFIX)
 
+# # get IssueTrackerSystem
+# aITS=myTestLink.getIssueTrackerSystem(aITSNAME)
+# print("getIssueTrackerSystem", aITS)    
+
 # Creates the project
 projInfo = 'Example created with Python %s API class %s in TL %s' % \
             ( myPyVersion, myApiVersion, myTLVersion )
 newProject = myTestLink.createTestProject(NEWPROJECT, NEWPREFIX, 
     notes=projInfo, active=1, public=1,
+#    itsname=ITSNAME, itsenabled=1,
     options={'requirementsEnabled' : 1, 'testPriorityEnabled' : 1, 
              'automationEnabled' : 1,  'inventoryEnabled' : 1})
 print("createTestProject", newProject)